use only truffle-config.js
